Refactor package dependencies and add react-intersection-observer
Update package-lock.json and package.json to include the new dependency react-intersection-observer. This will be used in the CompletedProjectCard component to handle intersection observer functionality for animations.

Refactor CompletedProjectCard component and use react-intersection-observer

In the CompletedProjectCard component, remove the framer-motion useInView hook and replace it with the react-intersection-observer useInView hook. This change allows for more control over the animation trigger and improves performance.

Update spring animation settings in CompletedProjectCard component

Adjust the spring animation settings in the CompletedProjectCard component to improve the smoothness and stability of the animation. This includes reducing stiffness and increasing damping.

Add animation reset and start value in CompletedProjectCard component

Before starting the animation in the CompletedProjectCard component, reset the value to 0 to ensure a smooth transition. Then, animate the value to the desired end value.

Refactor InfiniteCarousel component and add InfiniteBannerCarousel component

Make changes to the InfiniteCarousel component and introduce a new component called InfiniteBannerCarousel. This refactor improves the functionality and performance of the carousel feature.